refactor(metrics): extract on_time module
Move TripResult, classify_trip_delays, compute_route_daily, and their
tests out of the monolithic metrics/mod.rs into a dedicated on_time.rs.
